+++ bank-feeds/src/codat_bankfeeds/models/operations/update_company.py
@@ -7,8 +7,8 @@ from codat_bankfeeds.models.shared import (
 from codat_bankfeeds.types import BaseModel
 from codat_bankfeeds.utils import FieldMetadata, PathParamMetadata, RequestMetadata
 import pydantic
-from typing import Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class UpdateCompanyRequestTypedDict(TypedDict):