+++ bank-feeds/src/codat_bankfeeds/models/shared/pushoperation.py
@@ -14,8 +14,8 @@ from codat_bankfeeds.types import (
 )
 import pydantic
 from pydantic import model_serializer
-from typing import List, Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import List, Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class PushOperationTypedDict(TypedDict):