+++ bank-feeds/src/codat_bankfeeds/models/shared/sourceaccountwebhook.py
@@ -7,8 +7,8 @@ from .sourceaccountwebhookpayload import (
 )
 from codat_bankfeeds.types import BaseModel
 import pydantic
-from typing import Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class SourceAccountWebhookTypedDict(TypedDict):