+++ bank-feeds/src/codat_bankfeeds/models/shared/syncasbankfeeds.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 from .bankaccountoption import BankAccountOption, BankAccountOptionTypedDict
 from codat_bankfeeds.types import BaseModel
 import pydantic
-from typing import List, Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import List, Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class SyncAsBankFeedsTypedDict(TypedDict):