+++ bank-feeds/src/codat_bankfeeds/models/shared/syncconfiguration.py
@@ -5,8 +5,8 @@ from .syncasbankfeeds import SyncAsBankFeeds, SyncAsBankFeedsTypedDict
 from .syncasexpenses import SyncAsExpenses, SyncAsExpensesTypedDict
 from codat_bankfeeds.types import BaseModel
 import pydantic
-from typing import Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class SyncConfigurationTypedDict(TypedDict):